chore(aio): split eval report activity unit tests
Generated-By: PostHog Desktop
Task-Id: 855f1f55-d7f9-44ea-aec0-0ec12bdb8eb3
